data folders refactoring
