+++ src/store/index.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { plovaApi } from "./plova";
+import { plovaApi } from "./plova/plova.api";
 
 export const store = configureStore({
   reducer: {
